+++ CHANGELOG.md
@@ -137,3 +137,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - License: GNU AGPL-3.0
 
 [1.0.0]: https://github.com/chrilep/SimpleAI/releases/tag/v1.0.0
+[1.0.1]: https://github.com/chrilep/SimpleAI/releases/tag/v1.0.1